requirements.txt
# Core dependencies for SFT Training Pipeline
torch>=2.0.0
transformers>=4.30.0
datasets>=2.14.0
tokenizers>=0.13.0
accelerate>=0.20.0


# Data processing
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Configuration management
PyYAML>=6.0
dataclasses-json>=0.5.0

# Logging and monitoring
wandb>=0.15.0
tensorboard>=2.13.0

# Parameter-efficient fine-tuning (optional but recommended)
peft>=0.4.0
bitsandbytes>=0.39.0

# Development and debugging
tqdm>=4.65.0
colorama>=0.4.6

# Optional: For advanced optimizations
deepspeed>=0.9.0  # For large model training
apex  # NVIDIA Apex for mixed precision training (install from source)

# Optional: For specific model architectures
sentencepiece>=0.1.99  # For T5, ALBERT, XLNet tokenizers
protobuf>=3.20.0

# System utilities
psutil>=5.9.0  # For system monitoring
GPUtil>=1.4.0  # For GPU monitoring

# Testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0